Concordances can now be built faster using the forward indices.
- Forward index files structure improved for saving/loading speed
- Added inverse sort indices, case-insensitive sort order to FI terms
file
- Use ForwardIndex.open() to open a forward index. It detects the
version and instantiates the right class.
- Added the concept of a punctuation+whitespace forward index, called
"punct". It stores the characters between words and can thus be used to
construct concordances without having to refer to the content store
(which is slow because of disk cache limits and the need for
decompression). If "punct" forward index exists, BlackLab automatically
switches to using the forward indices to make concordances. Otherwise it
will keep using the content store. Note that concordances made from the
forward index do not include XML tags.