Optimized the way some queries are executed using new BLSpans base class.
BLSpans is the base class for all our Spans classes now, and includes
extra methods that can give guarantuees about properties of the hits
in a BLSpans object (such as: "all hits are of the same length"). These
are used to skip certain steps or to follow faster code paths for some
cases.