You can now choose case-sensitivity when sorting/grouping.
Internally, hit context is now stored in token IDs instead of in sort
order like before. Sort order is only looked up if/when needed.

As part of this, to keep the number of overloaded methods down a
little, some superfluous methods have been deprecated. Alternatives are
provided in the Javadoc.