Made HitsWindow copy Hit objects.
Before it referenced them, but that can cause problems when overwriting
the context[] array of the objects (since it belong to two different
collections, one Hits and one HitsWindow).

Hit is now cloneable and HitsWindow clones the Hits it's interested in.
This shouldn't be a problem as HitsWindow is usually used with small
(max. ca. 100) numbers of hits.