HitPropValue serialization now uses words instead of ids.
HitPropValueContextWord(s) used to include term ids in the
serialized string. If used in an URL (such as a URL for viewing one
group after grouping results), these cause the URL to not
be bookmarkable, because the term id might change if the data is
re-indexed. Now, we use the word in the serialized value, so reindexing
doesn't invalidate it.