Changed Searcher instantiation to a static open() method.
Search.open(File) takes the place of the now deprecated new
Searcher(File). The static factory method makes it possible to
change Searcher to an abstract class in the future and move the
implementation to a new SearcherImpl class.

This is useful for unit testing: many methods require a Searcher
object to be passed, but we don't always want an actual Searcher
with the required index on disk. Instead we want to provide a stub
Searcher object. Making Searcher an abstract class and moving the
implementation to a new class makes it easy to create such stubs as
needed.