Hacked in experimental SRU CQL support.
WARNING: this functionality is far from complete, bug-free or working as
it should be! It is just a quick and dirty starting off point for people
interested in developing this further. If you do look at it, let me know
what you think. I'd be happy to help if you want to work on this.

You can call ContextualQueryLanguageParser.parse() to get a
CompleteQuery object (containing both a TextPattern and a filter Query).
Then wrap the Query in a QueryWrapperFilter and call
Searcher.find(TextPattern, Filter). (This will obviously be streamlined
in the future)