Added an example directory with the QueryToolServlet application.
This application can be used as a debugging tool, and shows how to
create a BlackLab web application that supports multiple users querying
at the same time through a single Searcher.

Please note that this is not a typical web search application; it is in
fact a web frontend to a command-driven console application
(nl.inl.blacklab.tools.QueryTool). As such it doesn't have bookmarkable
URLs, etc. A future example will show how to set up that kind of web
application.