Made sure values are desensitized depending on the field.
When searching on a property that's indexed with the "insensitive only"
setting, but you're searching in "sensitive" mode, values will still be
desensitized for this property. If you didn't do this, you would never
find anything unless you put all your queries in lowercase, which isn't
very useful. In effect, the way the field is indexed may override the
"sensitive" search mode now.