Improved on the default analyzer for metadata indexing and queries.
BLDutchAnalyzer is the new default analyzer, and it should be an
improvement for most other languages as well. The main reason for this
change is that we don't want a dash (-) to be seen as a token end
(e.g. "ex-wife" should be a single token, not two). There's a few more
extra rules; see the comments for details.