You can now index metadata from external sources (e.g. a webservice).
The metadata fetcher system allows you to read the metadata for a
document from the filesystem, via HTTP, etc. To do so, implement a
MetadataFetcher class (see DocIndexerXmlHandlers.MetadataFetcher and the
two examples in nl.inl.blacklab.indexers) and start IndexTool with an
indexer.properties file containing a property "metadataFetcherClass"
that is the fully qualified class name of your MetadataFetcher class.
The MetadataFetcher class looks at the Lucene Document built from the
content, fetches external metadata (e.g. based on an id in the Lucene
document) and adds the metadata to the Lucene document.

This is a work in progress; it will be streamlined a bit more
in the future (e.g. prevent MetadataFetcher from being instantiated for
every document), but it works for now.