Slight change to how main property of complex fields is detected.
Main property is detected by looking for the property that has character
offsets, but not all indices may have those (right now they do, but in
the future they might not). In that case, the first property is assumed
to be the main property. This may not always be correct, so we may need
to improve how index structure is detected in the future.