Specifying the wrong format to IndexTool now results in warnings.
A warning is issued if a file was parsed but wasn't indexed as a
document; this usually means the root element for the specified format
was not found, i.e. it's a different format than specified.

At some future point, we'd like to try to autodetect formats, but there
are some performance concerns with that.