Added Kwic methods to get better-structured information for KWIC views.
The Kwic methods replace the Concordance methods for retrieving
keyword-in-context data for hits. The older Concordance methods
return XML snippets, which is not what you want in every circumstance
(i.e. implementing a JSON webservice). The new Kwic methods (i.e.
Hits.getKwic(hit)) return an instance of class Kwic, which contains the
context information in the form of String Lists. The old methods have
been reimplemented in terms of the new Kwic system, and Kwic objects
may be converted to Concordance objects if needed. However, the
Concordance parts may be deprecated in the future to reduce API
complexity and redundancy.