Moved context from Hit to Hits; keep Hits sort order in Integer array.
The object of these changes is to make the Hit class immutable so instances can be reused. The 'hits' instance variable of the Hits class could now be shared between multiple Hits instances with different sorts, contexts, etc.

As a result of this change, the HitProperty classes now all take a Hits object (the Hits object you're going to sort/group) and internally compare Hit objects based on their "original" index (the order in which they were added to Hits). Other than that, no changes to your code should be necessary.