Added JSON-reading/writing code.
We will use JSON to store index metadata, such as what BlackLab version
was used to create it, the types of its metadata fields, etc. (any
relevant information not stored in the Lucene index).

The code is the reference implementation by Douglas Crockford from
json.org, but it is placed in the package nl.inl.util.json to avoid any
potential conflicts. The class nl.inl.util.Json provides a few
convenience methods for dealing with commented-JSON files.