Added example of new indexmetadata.json file (not in use yet).
This file functions as both configuration for the indexer and metadata
for the BlackLab index. This allows us to store information about the
index that's not contained in the Lucene index itself, for example: what
analyzers were used for each field, display names for fields and the
build date of the BlackLab JAR used.

Also added a stub for the IndexMetadata class that will read and write
this file.