Added a way to add property values at any position while indexing.
This is useful for some corpus formats that refer back to earlier
parts of the document. E.g. FoLiA can specify named entities by
referring back to the words that comprise the entity. In this case,
it is convenient to add an inline tag at this position while indexing.
ComplexFieldProperty.addValueAtPosition(value, position) allows this.