Added "remove" strategy for fixing well-formedness of snippets.
Snippets and concordances might not be well-formed XML because they are
cut from a larger XML document. To fix this, we used to always add tags
to the start or end of the snippet. This might produce strange results
(i.e. sentence start tags mid-sentence). The new "remove" strategy might
be a better option in this case.

Theoretically, we could enhance the system in the future by providing
per-tag strategies, but this seems overkill for now.