Added DocContents classes and used them in Kwic and Concordance.
DocContents is an abstract base class with two child classes,
DocContentsString and DocContentsFromForwardIndex. They represent a part
of the document contents, either from the original input data or
reconstructed from the forward index.

Kwic and Concordance now use these to store the part of the content they
work with. We also plan to use them to implement generic "get part of
the document content" methods.