Eliminated SlowCompositeReaderWrapper in Hits.
SlowCompositeReaderWrapper turned out to be the cause of a strange
problem where rare terms were not found even though they occurred in the
index.

Instead of using SlowCompositeReaderWrapper, we now manually loop
through the index segments while gathering Hits. This works properly and
may be faster to boot.

SlowCompositeReaderWrapper is still used in a few other places, but it
appears to work correctly there; we will eventually probably follow the
same approach there as well.