Fixed dummy token at end of document sometimes being matched.
BlackLab used to sometimes add a "dummy" token at the end of a document
when (close) tags occurred after the last word. These dummy tokens would
sometimes be matched, e.g. by a matchall []. We couldn't fix this
because the dummy token would only sometimes be there, and we had no
quick way of knowing which documents had them and which didn't. Now we
simply index a dummy token at the end of each document and make sure we
never match the last token.