Added contentViewable metadata setting.
This setting indicates whether or not all end users may retrieve the
full contents of documents. Note that BlackLab itself does not use this
setting; it exists merely to provide a simple way for applications to
implement "all or nothing" access controls. BlackLab Server respects
this setting and will refuse to return the full contents of a document
if the setting is missing or set to false for an index.