Added default unknown condition and value to indextemplate.json.
This allows you to control what happens when a metadata value is
unknown, either because it's missing, empty, or both. You could
previously specify what should be done for each metadata field, but
because you don't always know in advance what metadata fields you will
encounter, and because it's convenient to be able to specify the desired
default behaviour once, you can now put "unknownCondition" and
"unknownValue" options under the "fieldInfo" object in
indextemplate.json. See "Indexing with BlackLab" on the project site for
more info.