Added HitPropertyContextWords.
This allows us to sort/group on more complex sequences of context words,
such as the first and the last word of the hit, or the second word to
the left of the hit. It can be a bit slower than HitPropertyHitText et
al, so only use this if you need the advanced functionality.