Deprecated TextPattern.toString() methods that take arguments.
toString() shouldn't take any outside information, but should just
describe the state of the object. The old situation caused problems not
just with testing, but with BLS's Job.Description as well. More in
general, you should be able to create and manipulate a TextPattern
without having a Searcher or QueryExecutionContext object, as it's just
an abstract query description.