Gave all Job* classes an inner JobDesc* class.
The point of this is to relieve Jobs of the responsibility of carrying
out their "input job" as well. So a sort job should ONLY sort, and not
be responsible for finding the hits to sort as well. The JobDesc classes
contain JobDescriptions of the input job, that can be executed (or
retrieved from cache) through SearchManager without having to know the
details of the job.