Rolled back attempt at constructing HitProperties without a Hits class.
We need Hits while constructing HitProperty* because it contains
settings such as the concordance field to use. This does mean we need to
keep sort/group criteria as strings until we have the Hits object we
want to apply them to. This can be solved, but is a bit more involved,
so we'll postpone it.