Upgraded from Lucene 5.2.1 to 5.3.0.
This involved wrapping the getSpans() and extractTerms() methods in a
SpanWeight class and adding a createWeight() method to the SpanQuery
class, and adding width() and collect() methods to the Spans classes,
among other changes.

As part of this upgrade, we've introduces BLSpanQuery, a base class for
(almost) all our SpanQuery classes. The exceptions are BLSpanTermQuery,
BLSpanOrQuery and BLMultiTermQueryWrapper, which derive directly from
their Lucene counterparts.