Extracted forward index id lookup code.
This uses DocValues or stored field. DocValues is currently used via
SlowCompositeReaderWrapper and UninvertingReader, which is slow and is
deprecated in Lucene 6.0. Now that this code is extracted to a neat
class, we can refactor it to use real DocValues (which will require
reindexing, so we might need to support te old method as well for a
while).