Don't build termIndex in search mode, saving memory and startup time.
The termIndex is used to find the term index for a term string (e.g. a
word). While searching, this is really only done when deserializing a
sort/group criterium (such as "cwo:word:s:that"), which is not a
time-critical operation. Instead of the termIndex, we use a binary
search through the memory-mapped terms file, which is plenty fast for
this usage and saves memory and some startup time.