Improved configuration error handling, made index scanning recursive.
If blacklab-server.json can't be found, BLS will return an XML error
message listing the search locations and a minimal configuration file.

Indices may now be in a subdirectory of one of your indexCollections,
and they will still be automatically discovered. If the directory is
named "index", the parent directory's name will be used for the index
name.