Added experimental support for enumerating subproperties.
Subproperties are a way to index e.g. different features of part of
speech in a single field. See DocIndexerOpenSonar for an example. Now,
if you've indexed subproperties, you can enumerate them in BlackLab
Server by going to /bls/indexName/fields/contents?subprops=pos to view
subproperties in the "pos" property of the "contents" field.