Made NFA matching exclusive to version 5 forward indices.
Because the index stores the sort order of the terms, which is
collator-based, we need to use the same collator while searching. If the
forward index version is 5 or greater, we will now use the collator that
doesn't ignore dash or space (which is how term comparison works
everywhere else), otherwise we will use the old one that does ignore
dash and space.

We switch off NFA matching for index versions older than 5, because
otherwise these might produce slightly different results with NFA
matching than with reverse matching (e.g. querying "sidestreet" normally
doesn't match "side-street", but with these older indices it would -
sometimes). If you want the speed increase of NFA matching, you should
re-index.