Fix index going into invalid state when uploading empty zip/tar
Some data about the fields in an index is not available until that index
contains some actual data. This data is attempted to be read out when
the index is not marked as empty. When uploading a tar/zip without
indexable documents, the index empty flag was cleared without some
actual data being present. This caused an exception to be thrown on
opening the index. 

Also fix a minor bug where requesting an invalid document id would throw
an internal server error instead of a sensible error message.