Refactor FileProcessor and TarGzipReader
- Removed duplicate code for handling tars/gzips/zips that was present
in FileProcessor largely into TarGzipReader. 

All references to indexing have been removed from the FileProcessor, it
now clearly reflects that it simply walks directories and archives. 
If useThreads is enabled, the FileHandler:stream function will be called
from separate threads. 

- TarGzipReader now can also process .zip files.
- TarGzipReader now closes the stream.
- TarGzipReader.FileHandler:handle will be called with a separate stream
from the input stream, so that it can safely be handled in other threads
and will never be prematurely closed by the TarGzipReader.