Extensive changes; slightly overhaul indexes in blacklab-server
- Added a new Index class to blacklab-server. The Index class manages
underlying Searcher and Indexer instances into the index. The Searcher
is only available while there is no Indexer opened on the Index, and
only a single Indexer can exist on an Index at a time. 
- IndexManager has been overhauled to use the new Index class. 
- RequestHandlerServerInfo and RequestHandlerIndexStatus now report the
progress of indexing if a certain Index is being indexed into at the
time.
- Fixed a bug where a handle to SearcherImpl would be kept open by the
searcherFromIndexReader map, which caused various resources and file
handles to remain open even after calling .close() on the Searcher. 
- Threaded indexing is tentatively enabled in blacklab-server. The
internal FileProcessor used to unpack to-be-indexed directories and
archives now spawns threads and calls the file handler in multiple
threads simultaeously. IndexTask and IndexThread have been removed.
- 

Issues
---------
- The pidfield used to be set from the IndexParam class. This is
currently not done. Seems to work but might have unforseen consequences. 
- Directory scanning for picking up placed/deleted indices could be a
little more refined.