Throw exception on missing format when indexing with format autodetect
The IndexStructure can contain a documentFormat string, which refers to
the configuration file that defines how new documents should be parsed
and indexed. When no DocIndexerFactory (responsible for creating and
configuring the actual DocIndexer that implements the indexing step) is
provided to the Indexer, this file is used to dynamically create the
DocIndexerFactory. 

If this file is deleted, and a factory was not provided, it's impossible
to determine how to actually index the documents, so now we throw an
exception straight away when this happens.