Initial DocumentFormat refactor
"format" here means the abstract concept of a DocIndexer with certain
parameters (the same DocIndexer class can potentially implement many
different formats, as is the case with the DocIndexerConfig).
"config" here means a json/yaml file used to configure a
DocIndexerConfig.

Everything generally works the same (a format has a name, is implemented
by a DocIndexer, and the Indexer can use the name to create the proper
DocIndexer for the job), but the concept of a format has been formalized
a little more.

The DocumentFormats class is still the central repository for all
formats that blacklab supports,
but it no longer directly manages the lists of DocIndexer and
ConfigInputFormat objects.
Instead, this task now falls to implementations of DocIndexerFactory.
The two core implementations are DocIndexerFactoryClass for the legacy
hardcoded DocIndexer types (the DocIndexerAbstract subclasses), and
DocIndexerFactoryConfig for the DocIndexerConfig implementations.

A "formatIdentifier" is just a unique string that names a specific
implementation and configuration of a DocIndexer (the basic
formatIdentifiers are tei, folia, csv/tsv, etc).

It's possible for users (as in programs using BlackLab) to register
their own factories, which can then expose new formats. These formats
can then be passed to the Indexer, and your custom factory will be able
to construct your customized DocIndexer.

------------------

BlackLab-Server uses this mechanism to register a factory that
specifically manages user's config formats. See
DocIndexerFactoryUserFormat. 
This way (once implemented), when a format is requested that is
identified as a custom user config (for instance by detecting the colon
separator) the factory can dynamically load the config and make it
available.