Added the option to set text direction of corpus.
The option textDirection can occur at the top-level of the indexmetadata file (values e.g. ltr/rtl). They can also be included in the corpusConfig block of the index format configuration file, in which case they will automatically be copied to the index metadata when creating a new index with this format.

BlackLab itself doesn't do anything with this option; the frontend can look at it and decide how to display results.