Fixed bug where counting total number of hits would never finish.
This was the cause of a semi-frequent problem when searching in the
WhiteLab2 application, where occasionally, searches would never finish.

This was caused by the fact that what we cache are Jobs, not Hits
instances, and related Jobs are linked using a reference count while
executing, but not after. Combined with what seems to be incorrect cache
management and the way the WhiteLab application talks to BLS, the
JobHits and JobHitsWindow instances for a new search were sometimes
quickly removed from the cache but JobHitsTotal was not, then the same
search was repeated, but JobHits and JobHitsWindow got a new Hits
instance while JobHitsTotal was still in the cache and using the old
Hits instance.

Long story short: this issue is now fixed, but it clearly shows that we
should be caching and re-using Hits instances instead of (or in addition
to?) Jobs instances. This will be a future refactoring.