Fixed "no value" vs "empty string" HitPropValue serialization issue.
E.g. when grouping on annotation on the word left of the hit, that
annotation can sometimes have the empty string as its value. Other times
the hit is the first token in the document, so the annotation will have
an invalid ("null") value. Both used to be were incorrectly serialized
to the same value; this has now been corrected.