Fix Indexer closing prematurely when uploading >1 file.
Also no longer throw exceptions before reading all uploaded data.
This was/is because browsers do not process the response body if the request
was aborted prematurely.
The server needs to finish reading the entire request before sending any
error codes or the user won't be able to see them.