Fix for linkedDocument storing.
If you specify linkedDocument.store = true, the whole linked document is
stored separately (not associated with a complex field). If you omit the
store setting or set it to false, the linked document indexer behaves as
normal (i.e. complex field in the linked document, if any, store
document content)