Added AuthDebugCookie for debugging logins.
Note that this turns out to be of limited use for our corpus-frontend,
as not all requests to blacklab-server originate from the client, so the
cookie is not always passed. The same is true for AuthHttpBasic.

Perhaps in the future we will update corpus-frontend to perform all
requests to blacklab-server from the client.