XSLT rule, displayAs option for each inline tag.
In the generated XSLT, each inline tag gets a rule that converts it to a
span with a CSS class generated from the xpath expression. To override
the CSS class, specify the displayAs option with the inlineTag
definition in the format configuration.

There are several predefined CSS classes such as sentence, paragraph,
line-beginning and page-beginning, and additional ones can be added as
needed.

(the explicit rule that was generated only for <p/> elements has been
taken out)