Improve plugin initialization system
- Added a new key "delayInitialization" in the "plugins" config,
If true, plugins will be initialized on first use instead of on load.
If false, plugins will immediately load, although this now happens in a
different thread to prevent locking up the main application.
When attempting to use a plugin that is being/hasn't been initialized, the operation blocks until
the initialization is complete.
- Don't remove plugins that failed to initialize, but throw an exception when trying to use them.
- Remove lazy-init code from TagPluginDutchTagger in favor of delayed init controlled by PluginManager.

Issue #101