Add support for "isVisible" flag in ConfigInputFormat.
Marking a format as hidden has no internal effects, but can be useful to
hide some advanced formats in client applications.

The CMDI config-based format is now marked hidden, because it only
indexes metadata and is a rather advanced use-case.

- Move reflection-based displayName, description, listFormat from
DocIndexer to DocIndexerAbstract (class-based DocIndexer
implementations) and rename listFormat function to isVisible.
- Ensure DocIndexerFactoryClass only uses DocIndexers that are subtypes
of DocIndexerAbstract.
- Rename unlisted boolean to isVisible in Format class.