Fix some thread blocking bugs when retrieving hits and documents.
Hits were accidentally always retrieved in blocking mode for the totals
count. This made queries returning large result sets extremely slow.
Also fixed a bug where requests for a single page of results shared a
lock with the count for the total results. This caused those requests to
wait for the totals count to finish, which could take a long time. The
HitsImpl and DocResults objects are now more aware of their threaded
nature and will internally poll instead of straight up block when this
happens.