Don't write valueListComplete metadata when state actually unknown
Fixes a bug where valueListComplete was always false for all metadata
properties explicitly named in the index's ConfigInputFormat.

When the initial indexMetadata is written while the index is created,
all known metadata fields are written out, and isValueListComplete() is
called on the (still empty) metadata field. This causes the valuelist to
written as empty. Then when the index is opened for writing, the
metadata file is read and set valueListComplete to false for those
metadata fields. Once this happens the boolean can never become true
again.