Changed generated XSLT from input format.
Whether or not a document has words in it is now decided using the same
expression as the word template. This means we now have a simple if/else
construction to decide whether the document has words, and we avoid
the problem we had before if the document xpath matches the root element.