Strip namespaces from document in generated xslt
This fixes a longstanding bug where namespaces could be ignored during
indexing, but then prevent words from being found when trying to render
the article as html. If no namespaces are provided in the
document/input/config format the whole pipeline should just ignore
namespaces. Previously this didn't happen, causing confusing issues
where documents were indexed fine but could not be displayed by the
corpus-frontend. Simply ignoring namespaces altogether in xslt is not(?)
possible, but we can preprocess the document to remove all namespaces
when the input format doesn't declare any namespaces. This should be a
safe operation, as when the config doesn't declare namespaces, the
xpaths in the config also cannot use namespaces, so removing any
namespaces that do happen to exist in the document itself should have no
influence.

Issue inl/corpus-frontend#134