Moved sorting hits to HitProperty, as a first step.
The goal is to eventually make HitProperty responsible for choosing how
to sort/group/filter for maximum performance. This way we can vary
implementations per property, for example properties that need context
vs. properties that don't.

One remaining issue is that we'd like to make HitProperty compare Hit
objects instead of original hit indexes like it does now, but we need
the original hit index to get the context, because that's pre-fetched,
keyed by original hit index. If fetching context for a single hit were
less expensive, it would be easier to move to comparing Hit objects
directly. Also helps lazily filtering hits for the same reason.