BLS: removed block parameter. Fixed wrongly async search jobs.
The block parameter allowed asynchronous searching with BlackLab, but
this was rarely used (it would be most useful for web applications, but
asynchronously handling responses from BLS is usually much preferable to
using block=no and having to poll to get the query status.

The totals count is asynchronous (unless you specify waitfortotal), and
this remains unchanged, because we want to give regular updates for hits
found so far.

In RequestHandlerDocs and RequestHandlerDocsGrouped, searchMan.search()
was called with block=false for the originalHitsSearch. But this is a
regular hits search, not a "total hits" search, so blocking is fine
(this just means that the Hits object is available, not that all hits
have been read).