Ensure there's a (dummy) annotated field for storing linked documents.
If you have metadata in a linked document, and want to store that linked
document, it has to be stored in a content store, which is attached to
an annotated field. In this case, you don't actually want any
annotations, so we create a dummy annotated field without any
annotations but with a content store.