Made maxMetadataValuesToStore configurable.
This affects how many metadata field values are stores in
indexmetadata.yaml, which in turn affects what values are shown in a
dropdown selection in the frontend (if the value list is complete,
dropdown is the default widget).

To configure, add a setting indexing.maxMetadataValuesToStore in your
config file and re-index your corpus.

e.g. in YAML:

indexing:
    maxMetadataValuesToStore: 100