Sped up determining relative frequency when grouping on metadata field.
A few flaws remain. If the group identity is empty, we can't search for
it (be sure to configure an "unknown value" before indexing to deal with
this). Right now we never tokenize the group identity, even for metadata
fields that are tokenized. Lastly, we don't use the specific analyzer
configured for the index or field yet.