Improved relative frequency reporting when grouping hits.
Now works correctly when grouping on untokenized fields as well. Now
takes the field analyzer into account.

When grouping on metadata field, if group identity is the empty string,
will now return 0 for the relative frequency instead of a nonsensical
number. (we can't search for the empty string, so we can't determine the
relative frequency in this case - configure an "unknown value" to avoid
this problem)