When requesting part of a document, make sure namespace prefixes are declared.
When requesting part of a document, a new document element is generated, because the original document element may not be included in the snippet. The newly generated document element should include any namespace prefix declarations that are used in the snippet.

NOTE: it is assumed that all namespace declarations are on the document element; if this is not the case, an exception will be thrown.