Report subcorpus size when grouping instead of relativeFrequency.
When grouping hits on (at least) a metadata field, subcorpus size will
be reported per group. If you're only grouping on context and not on
metadata, subcorpus size will be reported once in the summary.

When grouping docs, only the number of documents in the subcorpus will
be reported. (This may be expanded in the future to include the number
of tokens if there's demand for it)