Return subcorpus size of 0 if size cannot be determined.
For certain group identities, notably the empty string for a metadata
field, we cannot determine the subcorpus size (we can't find documents
where a metadata field is empty using Lucene). In this case, 0 is
returned. The application should detect this and deal with it. To avoid
the problem, configure an unknown condition and value in your input
format configuration file.