Made the number of threads used while indexing configurable.
More threads is faster but uses more memory (potentially many GBs for
large indexes). The default is now 2 threads.

In blacklab-server.yaml or blacklab.yaml, add the numberOfThreads option
to the indexing section to specify the number of threads.

For IndexTool, add the --threads 4 option to use 4 threads. This option
overrides the numberOfThreads option in blacklab.yaml.

The number of indexing threads is capped at at least 1 and at most the
number of CPU cores minus one.