Fixed occasionally missing valid matches due to an optimization bug.
Reported by @severian. Certain negative clauses (e.g. [word != 'the']+)
will sometimes be rewritten to a "position filter" clause, essentially
saying "i'm looking for n-grams of length 1 or more that DON'T contain
'the' ". This is a more efficient way of matching than finding all
tokens that are not 'the', but it does require us to know the document
length (or we couldn't generate all n-grams). The bug produced an
incorrect document length, causing us to miss valid matches.