Attempt to fix concurrency bug.
See #99, reported by @severian. NumericDocValues should only be used
from a single thread, but we violate this requirement when accessing the
forward index from multiple threads simultaneously, e.g. in
HitsFromQueryParallel.

We've attempted to fix the issue by synchronizing
DocIntFieldGetter.getFieldValue(). This does seem to reduce the
frequency of the problem, but doesn't solve it. We should actually
ensure NumericDocValues instances are only ever used from a single
thread.