When displaying left context, ensure the words are in original order.
Sorting/grouping on left context order the words in reverse, so the
rightmost word is the most significant while sorting. But while
displaying, we want the original order of the words.

The only situation in which this doesn't work as expected yet is if you
sort/group on multiple parts, some of which are reverse direction.