Initial implementation of multiple metadata values for documents
Adds a new "multipleValues" option to metadata definitions in config
files.
To stay backwards-compatible, a new query parameter
"multiplevalues=true" must be passed to /docs/docId and result requests
(/docs and /hits). Doing this will return metadata as a json array
instead of a string in the case of json, and as
<metadataFieldName><value>value</value></metadataFieldName> in the case
of xml.