Made titleField detection stricter.
If no titleField specified, but a field named "title" (or a variation
with different case such as "Title") exists, it is assumed to be the
titleField. Otherwise, fromInputFile (the name of the file the document
came from) is used.

Before we just looked for fields containing the word "title", but this
can lead to unexpected and undesirable outcomes.