Also map values when indexing a regular multiple-values metadata field.
Mapping was done for forEach fields in all cases and for regular fields
if it was not a multiple-values field. Should always be done.

Question: if a field is indexed using forEach, should process and/or
mapping from the regular field definition (if it exists) apply as well
or not? (right now, it does not, but this seems wrong)