Made warning about longer metadata values clearer.
Values longer than 256 are indexed/stored in Lucene as normal, but
they're not stored in indexmetadata.yaml (which we use to enumerate all
values of a field, e.g. to display a selection list).

Eventually we'll move to using DocValues instead of storing the values
in indexmetadata.yaml and this difference will go away.