Log a message to indicate where SQLite library will likely be extracted.
Only applies if detailed logging to SQLite is configured (default off).
If that's the case, SQLite will extract a native library to a temporary
directory. This is usually java.io.tmpdir, so we log that value. Make
sure the current user is allowed to write there and nothing is blocking
execution of code in that location (e.g. noexec mount flag, SELinux).