Added debug auth method via HTTP headers.
This is an insecure way of authenticating to BlackLab Server by sending
two HTTP headers. It is only intended for testing purposes.

First, in the authentication section of blacklab-server.yaml, add the
debugHttpHeaderAuthToken parameter. The value of this will be your
'secret' access token.

Next, send your requests to BlackLab Server with the extra HTTP headers
X-BlackLabAccessToken (the 'secret' value from above) and
X-BlackLabUserId (the user you wish to authenticate as).

Needless to say this method is insecure because it allows full access to
all users' corpora, and the access token could potentially leak to an
attacker.