Fix sorting & grouping on document metadata with multiple values
This fix contains 3 parts:
- The metadata values being stored in the lucene docs need to be sorted,
so that sensible comparisons can be made across documents. A document
with values [a,b] for a field should be equal to a document with values
[b,a] for that same field. This is only the case if these values are in
the same order for both documents. This is the easy part, a simple
change in DocIndexer is enough.
- All values must be considered when comparing documents (to find the
appropriate group or sorting order). This is done by modifying the
DocPropertyStoredField to return a PropValueMultiple instead of
PropValueString.
- Lastly, a PropValueMultiple with only one internal PropValue should
compare equal (and hash to the same number as) to that single PropValue
(PropValueString, PropValueNumber, etc.). This is required because when
blacklab-server retrieves a request to list the documents in a group "a"
for instance, it parses that group identifier "a" to a PropValueString,
which is then used to find the right group. The group however has a
PropValueMultiple as main identifier.