Enable/fix sorting & grouping on document metadata with multiple values
To allow this two changes were necessary:
1. All values must be considered when comparing documents (to find the
appropriate group or sorting order). This is done by modifying the
DocPropertyStoredField to return a PropValueMultiple instead of
PropValueString.

2. An option needed to be added to allow sorting a metadata field's values during indexing (default false).
When a metadata field has multiple values, the order is preserved by default.
This can sometimes cause issues, as documents with (a,b,c) will be sorted into a different group than documents with values (a,c,b).
Sometimes this is what you want (for example with regional codings, or a series of dates, or authors of newspapers), but other times, the order is important (authors of scientific papers).
So sorting can now be enabled, usage as follows (.yaml.blf file).

metadata:
  fields:
  - name: myMetadataField
    sortValues: true

During document retrieval/sorting/grouping, we've opted to concatenate all values into a single PropertyValueString for performance reasons.
This has the side effect that when creating the group's displaynames, a seperator needed to be decided.
For this we've opted for the middle dot '·' as it's commonly recognised as a separator, but is also relatively rare to encounted otherwise.
That way users can be mostly certain that it's a separator and not a part of one of the individual values.