Add option to set sorting per metadata field (default false)
Usage:
metadata:
  fields:
  - name: myMetadataField
    sortValues: true

Changing this influences how documents are sorted and grouped among each other.
For example, two documents [a,b] and [b,a] will normally be grouped separately, because the arrays of values
are not strictly equal.
Enabling sorting of values however, will store the values for both documents as [a,b], causing them to be lumped into the same group later.
We've left this as an option because sorting is not always what you want - image for instance a metadata field list a paper's authors, the order would be important!

For normal usage however, setting sortValues usually will result in a more intuitive ordering.