Fix issue with retrieving group contents
When grouping documents without value for a docprop, they are grouped in the "null" or "" (empty string) group.
When this was sent to the client previously, this group's identifier was "str:", indicating empty string as value.
However, when using the DocPropertyStoredValue class, it was generated as a PropertyMultiple that serialized to just "" (empty string).
When this was sent back to BlackLab as viewGroup parameter, it would be discarded, causing the normal groups to be returned.

In order to solve this, the group's identifier once again needs to serialize to "str:" (empty PropertyValueString).
There are two ways to do this:
- in DocPropertyStoredField, return a PropertyValueMultiple containing a single PropertyValueString encoding the empty string
- in DocPropertyStoredField, return a regular DocPropertyString for when the document has no value here.

The second option has the caveat that the compare functions between PropertyValueString and PropertyValueMultiple need to be compatible, as we'll be sorting arrays based on them.
This is slightly more work, but seems like the more robust approach.